+++ BranchSDK-Samples/console/storagedemo/storagedemo.cpp
@@ -1,5 +1,5 @@
 #include <BranchIO/Util/Log.h>
-#include <BranchIO/Storage.h>
+#include <BranchIO/Util/Storage.h>
 #include <Poco/Exception.h>
 #include <exception>
